Consolidate plugin normalization into single migration
Merge graphql/openapi/mcp/google-discovery normalization into one
migration per app: 0007_normalize_plugin_secret_refs (local sqlite)
and 0008_normalize_plugin_secret_refs (cloud pg, mcp-only since cloud
doesn't ship google-discovery).

Each test now imports a shared PRE_0007_SQL helper that seeds all four
plugins' parent tables at the pre-migration shape — needed because the
single consolidated migration touches every plugin in one transaction.
